Merge SMGCommunity/Petari at c600c594, favoring upstream conflicts.
Synchronize Game mirrors and dependencies while retaining the native
provider boundaries needed by the PC runtime. Use upstream BGM volume
and animation scale implementations and add portable collision/rumble data.

Adapt camera, talk, material, binder, math, and Player interfaces to the
merged declarations. Preserve existing locally recovered functions where
upstream remains incomplete. Record source dispositions and validation.

Validated both PC executables and nine focused regression targets on
macOS arm64, including real-disc Gateway movement and talk tests.
Pre-existing working changes are preserved outside this commit.

+++ pc-port/src/Game/Screen/LensFlare.hpp
@@ -97,5 +97,5 @@ namespace MR {
     void addBrightObj(BrightObjBase* brightObj);
     void removeBrightObj(BrightObjBase* brightObj) noexcept;
     void setLensFlareDrawSyncToken();
-    u16 getLensFlareDrawSyncTokenIndex();
+    u32 getLensFlareDrawSyncTokenIndex();
 }  // namespace MR
